Refactor project structure and make raw block exporter
We have set-up this repo to support multiple projects in
subfolders. We have also created a new project `raw-exporter` which is
a trimmed down version of the old exporter and which takes only raw
blocks and stores them in Kafka. The old exporter is still in the repo
in the `old-exporter` project.