Refactor transfers and stacks topics
Now we create three topics with three partitions each. The three
topics are created so that the three clickhouse servers can each
ingest data from one topic. Each topic has three partitions to make
sure that writing is balanced to the three Kafka brokers.

We have refactored the way migrations are done. To apply this change
on the kubernetes clusters we'll have to clean the Zookeeper migration
state.